Use testthat::test_dir() instead of devtools::test()
Since devtools::test() is meant for interactive use, it does not fail if
any test fails. A good alternative is provided by testthat::test_dir().
That does require the package to be installed, though. To ease the
installation, /usr/local/lib/R/site-library in the docker image is made
world-writeable.